feat: common, eureka init 설정
