+++ .github/pull_request_template.md
@@ -4,7 +4,7 @@
 
 ### PR 타입(하나 이상의 PR 타입을 선택해주세요)
 
--[ ] 신규 기능 추가/수정
+- [ ] 신규 기능 추가/수정
 - [ ] 버그 수정
 - [ ] 리팩토링
 - [ ] 설정
