+++ common/src/main/resources/applicatioin.yml
@@ -1,0 +1,3 @@
+spring:
+  application:
+    name: common
