+++ eureka/src/main/resources/application.yml
@@ -1,0 +1,17 @@
+spring:
+  application:
+    name: eureka
+
+server:
+  port: 19090
+
+eureka:
+  client:
+    service-url:
+      defaultZone: http://localhost:19090/eureka/ # ????? ??????? ??? ??? ???? ????? ?????? ??? ? ??? URL? ?????.
+    register-with-eureka: false  # ?? Eureka ??? ? ??? ???? ??
+    fetch-registry: false  # ?? Eureka ??? ?????? ???? ??
+  server:
+    enable-self-preservation: false  # ?? ?? ?? ????
+  instance:
+    hostname: localhost # ??? ??? ??? ??? ??? ?? ???? ?? ? ?????.
