+++ eureka/src/test/java/com/team15gijo/eureka/EurekaApplicationTests.java
@@ -1,0 +1,13 @@
+package com.team15gijo.eureka;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EurekaApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
