+++ user/src/main/java/com/team15gijo/user/UserApplication.java
@@ -1,0 +1,14 @@
+package com.team15gijo.user;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+
+@SpringBootApplication(scanBasePackages = {"com.team15gijo.user", "com.team15gijo.common"})
+public class UserApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(UserApplication.class, args);
+    }
+
+}
