+++ user/src/main/resources/application.yml
@@ -1,0 +1,13 @@
+spring:
+  application:
+    name: user-service
+
+server:
+  port: 19092
+
+eureka:
+  client:
+    register-with-eureka: true
+    fetch-registry: true
+    service-url:
+      defaultZone: http://localhost:19090/eureka/
