+++ user/src/test/java/com/team15gijo/user/UserApplicationTests.java
@@ -1,0 +1,13 @@
+package com.team15gijo.user;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
